Added stat service test

+++ src/main/java/com/test/mshmatov/dto/AverageSpeedStatDto.java
@@ -7,5 +7,5 @@ import lombok.*;
 @Builder
 public class AverageSpeedStatDto {
     private RunDto run;
-    private Float averageSpeed;
+    private Long averageSpeed;
 }
